requirements.txt
wandb
jax[tpu]
flax
pytorch
optax
tensorflow
